Moved Resources\ApiResponse to Models namespace
