+++ src/Mailgun/Model/Domain/CreateCredentialResponse.php
@@ -9,7 +9,7 @@
 
 namespace Mailgun\Model\Domain;
 
-use Mailgun\Resource\ApiResponse;
+use Mailgun\Model\ApiResponse;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
