+++ src/Mailgun/Model/Message/SendResponse.php
@@ -9,7 +9,7 @@
 
 namespace Mailgun\Model\Message;
 
-use Mailgun\Resource\ApiResponse;
+use Mailgun\Model\ApiResponse;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
