+++ src/Mailgun/Model/Routes/Response/ShowResponse.php
@@ -10,7 +10,7 @@
 namespace Mailgun\Model\Routes\Response;
 
 use Mailgun\Model\Routes\Route;
-use Mailgun\Resource\ApiResponse;
+use Mailgun\Model\ApiResponse;
 
 /**
  * @author David Garcia <me@davidgarcia.cat>
