+++ src/Mailgun/Model/Suppressions/Complaint/ShowResponse.php
@@ -9,7 +9,7 @@
 
 namespace Mailgun\Model\Suppressions\Complaint;
 
-use Mailgun\Resource\ApiResponse;
+use Mailgun\Model\ApiResponse;
 
 /**
  * @author Sean Johnson <sean@mailgun.com>
