+++ src/Mailgun/Model/Tag/UpdateResponse.php
@@ -9,7 +9,7 @@
 
 namespace Mailgun\Model\Tag;
 
-use Mailgun\Resource\ApiResponse;
+use Mailgun\Model\ApiResponse;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
